Moved create run dialog component into its own directory. Fixed unit tests.
