+++ src/main/webapp/site/src/app/teacher/create-run-dialog/create-run-dialog.component.html
@@ -1,16 +1,16 @@
 <h2>Run with class: {{project.name}} (Lesson ID: {{project.id}})</h2>
 <div>
-  <h3>1. Set Periods</h3>
+  <h3>1. Choose Periods</h3>
   <div fxLayout="row">
     <div fxFlex>
       <div *ngFor="let period of [1, 2, 3, 4]">
-        <mat-checkbox [(ngModel)]="periods[period]" (change)="periodChanged()">Period {{period}}</mat-checkbox>
+        <mat-checkbox [(ngModel)]="periods[period]" (change)="periodChanged()">{{period}}</mat-checkbox>
         <br/>
       </div>
     </div>
     <div fxFlex>
       <div *ngFor="let period of [5, 6, 7, 8]">
-        <mat-checkbox [(ngModel)]="periods[period]" (change)="periodChanged()">Period {{period}}</mat-checkbox>
+        <mat-checkbox [(ngModel)]="periods[period]" (change)="periodChanged()">{{period}}</mat-checkbox>
         <br/>
       </div>
     </div>
@@ -22,8 +22,9 @@
   </div>
   <div>(Separate manually named periods with commas. Example: "Section1, Section 2, Section 3". Manually named periods should be no more than 16 characters in length.)</div>
 </div>
+<br/>
 <div>
-  <h3>2. Set Students Per Team</h3>
+  <h3>2. Choose Students Per Team</h3>
   <p>How many students per computer during the lesson?</p>
   <mat-radio-group [(ngModel)]="studentsPerTeam">
     <mat-radio-button value="1">Only 1 student</mat-radio-button>
@@ -31,8 +32,9 @@
     <mat-radio-button value="3">1-3 students</mat-radio-button>
   </mat-radio-group>
 </div>
+<br/>
 <div>
-  <h3>3. Set Start Date</h3>
+  <h3>3. Choose Start Date</h3>
   <mat-form-field>
     <input [(ngModel)]="startDate" matInput [matDatepicker]="picker" placeholder="Choose a date" disabled>
     <mat-datepicker-toggle matSuffix [for]="picker"></mat-datepicker-toggle>
